Almost done

+++ app/Http/Controllers/Admin/PlanServiceCategoryController.php
@@ -1,8 +1,12 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
 use App\Models\PlanServiceCategory;
+use App\Models\PlanServiceCategorySupportingDocumentItem;
+use App\Models\PlanServiceOrder;
+use App\Models\SupportingDocumentCategory;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Routing\Controller;
@@ -18,7 +22,7 @@ class PlanServiceCategoryController extends Controller
      */
     public function index()
     {
-       return view('plan_category.index');
+       return view('admin.plan_category.index');
     }
     public function dataTable()
     {
@@ -27,7 +31,8 @@ class PlanServiceCategoryController extends Controller
             ->addIndexColumn()
             ->addColumn('action', function(PlanServiceCategory $planCategory) {
                 $btn = '';
-                $btn .= '<a href="' . route('plan-service-category.edit', ['plan_service_category' => $planCategory->id]) . '" class="btn btn-purple bg-gradient-purple btn-sm btn-edit"><i class="fa fa-edit"></i></a>';
+                $btn .= ' <a href="' . route('add-plan-service-category-supporting-document-items', ['planServiceCategory' => $planCategory->id]) . '" class="btn btn-purple bg-gradient-primary btn-sm"><i class="fa fa-folder"></i> Add document categories</a>';
+                $btn .= ' <a href="' . route('plan-service-category.edit', ['plan_service_category' => $planCategory->id]) . '" class="btn btn-purple bg-gradient-purple btn-sm btn-edit"><i class="fa fa-edit"></i></a>';
                 $btn .= ' <a role="button" data-id="' . $planCategory->id . '" class="btn btn-danger btn-sm btn-delete"><i class="fa fa-trash"></i></a>';
 
                 return $btn;
@@ -50,7 +55,7 @@ class PlanServiceCategoryController extends Controller
     {
         $maxSort = en2bn(PlanServiceCategory::max('sort') + 1);
 
-        return view('plan_category.create',compact('maxSort'));
+        return view('admin.plan_category.create',compact('maxSort'));
     }
 
     public function store(Request $request)
@@ -84,14 +89,15 @@ class PlanServiceCategoryController extends Controller
             // Roll back the transaction in case of an error
             DB::rollback();
             // Handle the error and redirect with an error message
-            return redirect()->route('plan-service-category.create')->with('error', 'An error occurred while creating the plan category : '.$e->getMessage());
+            return redirect()->route('plan-service-category.create')->withInput()
+                ->with('error', 'An error occurred while creating the plan category : '.$e->getMessage());
         }
     }
 
     public function edit(PlanServiceCategory $plan_service_category)
     {
 
-        return view('plan_category.edit',compact('plan_service_category'));
+        return view('admin.plan_category.edit',compact('plan_service_category'));
     }
 
     public function update(PlanServiceCategory $plan_service_category,Request $request)
@@ -134,6 +140,14 @@ class PlanServiceCategoryController extends Controller
     public function destroy(PlanServiceCategory $plan_service_category)
     {
         try {
+            $planServiceOrder = PlanServiceOrder::
+            where('plan_service_category_id',$plan_service_category->id)
+                ->first();
+
+            if ($planServiceOrder){
+                return response()->json(['success'=>false,'message' => "Can't delete, plan service category has service order"], Response::HTTP_OK);
+
+            }
             // Delete the User record
             $plan_service_category->delete();
             // Return a JSON success response
@@ -141,6 +155,48 @@ class PlanServiceCategoryController extends Controller
         } catch (\Exception $e) {
             // Handle any errors, such as record not found
             return response()->json(['success'=>false,'message' => 'Plan service category not found: '.$e->getMessage()], Response::HTTP_OK);
+
+        }
+    }
+
+    public function addSupportingDocumentItems(PlanServiceCategory $planServiceCategory)
+    {
+        $supportingDocumentCategories = SupportingDocumentCategory::orderBy('sort')->where('status',1)->get();
+        return view('admin.plan_category.add_document_categories',compact('planServiceCategory',
+        'supportingDocumentCategories'));
+    }
+    public function addSupportingDocumentItemsPost(PlanServiceCategory $planServiceCategory,Request $request)
+    {
+        // Start a database transaction
+        DB::beginTransaction();
+
+        try {
+            if ($request->status){
+                PlanServiceCategorySupportingDocumentItem::where('plan_service_category_id',$planServiceCategory->id)
+                            ->delete();
+
+                $counter = 0;
+                foreach ($request->status as $reqStatus){
+                    $supportingDocumentItem = new PlanServiceCategorySupportingDocumentItem();
+                    $supportingDocumentItem->plan_service_category_id = $planServiceCategory->id;
+                    $supportingDocumentItem->supporting_document_category_id = $request->status[$counter];
+                    $supportingDocumentItem->sort = $request->sort[$counter] ?? 1;
+                    $supportingDocumentItem->save();
+
+                    $counter++;
+                }
+            }
+
+            DB::commit();
+
+            // Redirect to the index page with a success message
+            return redirect()->route('plan-service-category.index')->with('success',"Plan service category's document items added successfully");
+        } catch (\Exception $e) {
+            // Roll back the transaction in case of an error
+            DB::rollback();
+
+            // Handle the error and redirect with an error message
+            return redirect()->route('add-plan-service-category-supporting-document-items',['planServiceCategory'=>$planServiceCategory->id])->with('error', 'An error occurred while updating the plan service category document item : '.$e->getMessage());
         }
     }
 }
